[BROOKLYN-183] Reuse classic init code, loading CAMP platform
