+++ karaf/features/src/main/feature/feature.xml
@@ -288,6 +288,7 @@
 
     <feature name="brooklyn-osgi-launcher" version="${project.version}" description="Brooklyn init">
         <feature>brooklyn-core</feature>
+        <feature>brooklyn-software-base</feature>
         <bundle>mvn:org.apache.brooklyn/brooklyn-launcher-common/${project.version}</bundle>
         <bundle>mvn:org.apache.brooklyn/brooklyn-karaf-init/${project.version}</bundle>
     </feature>
