+++ karaf/init/pom.xml
@@ -44,6 +44,7 @@
             <artifactId>brooklyn-launcher-common</artifactId>
             <version>${project.version}</version>
         </dependency>
+
         <dependency>
             <groupId>org.testng</groupId>
             <artifactId>testng</artifactId>
